+++ spec/bare_spec.rb
@@ -21,8 +21,8 @@ describe "CLI" do
       execute_script("bin/searchlink", use_bundler: true, stdin_data: input)
       expect(last_execution).to be_successful
 
-      result_markup = /#{Regexp.escape("[The Strata (Original Soundtrack), by Mark R. Healy]")}/
-      result_bare = /#{Regexp.escape("<https://itunes.apple.com/us/app/id325946571>")}/
+      result_markup = /#{Regexp.escape('[The Strata (Original Soundtrack), by Mark R. Healy]')}/
+      result_bare = /#{Regexp.escape('<https://itunes.apple.com/us/app/id325946571>')}/
       expect(last_execution.stdout).to match(result_markup)
       expect(last_execution.stdout).to match(result_bare)
     end