+++ spec/history_spec.rb
@@ -1,12 +1,12 @@
 # frozen_string_literal: true
 
-require 'spec_helper'
+require "spec_helper"
 
 # https://github.com/thoiberg/cli-test
-describe 'CLI' do
+describe "CLI" do
   include CliTest
 
-  return if ENV['SKIP_HISTORY'] == 'true'
+  return if ENV["SKIP_HISTORY"] == "true"
 
   # Array of searches and results
   # [[search, expected result],...]
@@ -21,17 +21,17 @@ describe 'CLI' do
     # ['* A Chrome history search [%](!hch boss begleri)', 'https://www.aroundsquare.com/collections/begleri/products/micro-boss-begleri-exotics'],
     # ['* An Edge bookmark search [%](!heb 2022 hardware)', 'https://brettterpstra.com/2022/12/31/bretts-favorites-2022-hardware'],
     # ['* An Edge history search [%](!heh brett terpstra)', 'https://brettterpstra.com/'],
-    ['* A Firefox bookmark search [%](!hfb 2022 hardware)', 'https://brettterpstra.com/2022/12/31/bretts-favorites-2022-hardware'],
-    ['* A Firefox history search [markdown exports mindmeister](!hfh)', 'https://gist.github.com/ttscoff/cd2a6c17964cccfb6665'],
-    ['* A Safari bookmark search [%](!hsb visionmedia mocha)', 'https://github.com/visionmedia/mocha'],
-    ['* A Safari history search [%](!hsh oracle template repo)', 'oracle/template-repo']
+    ["* A Firefox bookmark search [%](!hfb 2022 hardware)", "https://brettterpstra.com/2022/12/31/bretts-favorites-2022-hardware"],
+    ["* A Firefox history search [markdown exports mindmeister](!hfh)", "https://gist.github.com/ttscoff/cd2a6c17964cccfb6665"],
+    # ["* A Safari bookmark search [%](!hsb visionmedia mocha)", "https://github.com/visionmedia/mocha"],
+    ["* A Safari history search [%](!hsh oracle template repo)", "oracle/template-repo"],
   ]
 
   searches.each.with_index do |search, i|
-    describe 'executes search' do
+    describe "executes search" do
       context "(#{i + 1}/#{searches.count}) when given the text '#{search[0]}'" do
         it "returns #{search[1]}" do
-          execute_script('bin/searchlink', use_bundler: true, stdin_data: search[0])
+          execute_script("bin/searchlink", use_bundler: true, stdin_data: search[0])
           expect(last_execution).to be_successful
 
           result = search[1].is_a?(Regexp) ? search[1] : /#{Regexp.escape(search[1])}/